+++ app/src/main/java/com/criteo/testapp/StandaloneActivity.java
@@ -18,13 +18,11 @@ package com.criteo.testapp;
 
 import static com.criteo.testapp.PubSdkDemoApplication.INTERSTITIAL_IBV_DEMO;
 import static com.criteo.testapp.PubSdkDemoApplication.NATIVE;
-import static com.criteo.testapp.PubSdkDemoApplication.STANDALONE_BANNER;
 
-import android.content.Context;
 import android.os.Bundle;
 import android.util.Log;
 import android.widget.Button;
-import android.widget.LinearLayout;
+import android.widget.FrameLayout;
 import androidx.appcompat.app.AppCompatActivity;
 import com.criteo.publisher.CriteoBannerView;
 import com.criteo.publisher.CriteoInterstitial;
@@ -44,12 +42,11 @@ public class StandaloneActivity extends AppCompatActivity {
   private static final InterstitialAdUnit INTERSTITIAL = new InterstitialAdUnit(
       "/140800857/Endeavour_Interstitial_320x480");
 
-  private Context context;
-  private LinearLayout adLayout;
   private CriteoBannerView criteoBannerView;
   private CriteoNativeLoader nativeLoader;
   private Button btnShowInterstitial;
   private Button btnShowInterstitialIbv;
+  private FrameLayout nativeAdContainer;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
@@ -57,20 +54,27 @@ public class StandaloneActivity extends AppCompatActivity {
     setContentView(R.layout.activity_stand_alone);
     MockedIntegrationRegistry.force(Integration.STANDALONE);
 
-    adLayout = findViewById(R.id.AdLayout);
     btnShowInterstitial = findViewById(R.id.buttonStandAloneInterstitial);
     btnShowInterstitialIbv = findViewById(R.id.buttonStandAloneInterstitialIbv);
+    criteoBannerView = findViewById(R.id.criteoBannerView);
+    nativeAdContainer = findViewById(R.id.nativeAdContainer);
 
-    context = getApplicationContext();
+    criteoBannerView.setCriteoBannerAdListener(new TestAppBannerAdListener(
+        TAG, "Standalone"));
 
     nativeLoader = new CriteoNativeLoader(
         NATIVE,
-        new TestAppNativeAdListener(TAG, NATIVE.getAdUnitId(), adLayout),
+        new TestAppNativeAdListener(TAG, NATIVE.getAdUnitId(), nativeAdContainer),
         new TestAppNativeRenderer()
     );
 
-    findViewById(R.id.buttonStandAloneBanner).setOnClickListener(v -> loadBanner());
+    findViewById(R.id.buttonStandAloneBanner).setOnClickListener(v -> loadBannerAd());
     findViewById(R.id.buttonStandAloneNative).setOnClickListener(v -> loadNative());
+  }
+
+  private void loadBannerAd() {
+    Log.d(TAG, "Banner Requested");
+    criteoBannerView.loadAd();
   }
 
   @Override
@@ -94,19 +98,6 @@ public class StandaloneActivity extends AppCompatActivity {
     if (interstitial.isAdLoaded()) {
       interstitial.show();
     }
-  }
-
-  private void loadBanner() {
-    if (criteoBannerView != null) {
-      criteoBannerView.destroy();
-    }
-
-    criteoBannerView = new CriteoBannerView(context, STANDALONE_BANNER);
-    criteoBannerView.setCriteoBannerAdListener(new TestAppBannerAdListener(
-        TAG, "Standalone", adLayout));
-
-    Log.d(TAG, "Banner Requested");
-    criteoBannerView.loadAd();
   }
 
   private void loadNative() {