+++ app/src/main/java/com/criteo/testapp/StandaloneRecyclerViewActivity.kt
@@ -77,7 +77,7 @@ class StandaloneRecyclerViewActivity : AppCompatActivity() {
 
   private fun loadBanner(adapter: Adapter) {
     val bannerView = CriteoBannerView(baseContext, STANDALONE_BANNER)
-    bannerView.setCriteoBannerAdListener(object : TestAppBannerAdListener(javaClass.simpleName, "Banner", null) {
+    bannerView.setCriteoBannerAdListener(object : TestAppBannerAdListener(javaClass.simpleName, "Banner") {
       override fun onAdReceived(view: View?) {
         adapter.addBannerAd(bannerView)
       }